requirements.txt
ansible==2.9.6
apache-libcloud==2.8.0
argcomplete==1.8.1
attrs==19.3.0
auto-sklearn==0.12.6
Automat==0.8.0
blinker==1.4
boto3==1.16.56
botocore==1.19.56
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.1
cloudpickle==1.6.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
ConfigSpace==0.4.18
constantly==15.1.0
cryptography==2.8
Cython==0.29.23
dask==2021.5.0
dbus-python==1.2.16
distributed==2021.5.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
entrypoints==0.3
fsspec==2021.5.0
HeapDict==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-import==0.2.2
liac-arff==2.5.0
locket==0.2.1
lockfile==0.12.2
MarkupSafe==1.1.0
more-itertools==4.2.0
msgpack==1.0.2
netaddr==0.7.19
netifaces==0.10.4
ntlm-auth==1.1.0
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
partd==1.2.0
psutil==5.8.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.3.0
pynisher==0.6.4
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrfr==0.8.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
pywinrm==0.3.0
PyYAML==5.3.1
requests==2.22.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-unixsocket==0.2.0
s3transfer==0.3.4
scikit-learn==0.24.2
scipy==1.6.3
SecretStorage==2.3.1
selinux==3.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smac==0.13.1
sortedcontainers==2.4.0
ssh-import-id==5.10
systemd-python==234
tblib==1.7.0
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
xmltodict==0.12.0
zict==2.0.0
zipp==1.0.0
zope.interface==4.7.1
